HW 14: Scheduler v1 is ready

+++ hw12_13_14_15_calendar/migrations/20240310134139_init.sql
@@ -4,9 +4,9 @@ CREATE table events (
                        user_id         uuid,
                        title           text,
                        date_and_time   timestamp with time zone,
-                       duration        bigint,
+                       duration        int,
                        description     text,
-                       time_to_notify  bigint
+                       time_to_notify  int
 );
 
 -- +goose Down
